backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jatin Shukla",email = "jatinshukla.lek@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (>=4.0.1,<5.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
httpx = "^0.28.1"
black = "^24.1.0"
ruff = "^0.2.0"
mypy = "^1.6.0"
types-python-jose = "^3.5.0.20250531"
types-passlib = "^1.7.7.20260211"
types-redis = "^4.6.0.20241004"

[tool.black]
extend-exclude = 'alembic[\\/]+versions[\\/].*'

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
exclude = 'alembic[\\/]+versions[\\/].*'

[tool.pytest.ini_options]
pythonpath = ["."]
